shared/pyproject.toml
[project]
name = "alfred-shared"
version = "0.1.0"
description = "Shared auth and config utilities for the Alfred platform"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0,<1.0.0",
    "pydantic>=2.0.0,<3.0.0",
    "pydantic-settings>=2.8.0,<3.0.0",
    "python-jose[cryptography]>=3.3.0,<4.0.0",
]

[tool.setuptools]
py-modules = ["auth", "config"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
